Add AWS, South Korea and Japan (#76)
Added a picture of my desk when I was working at AWS in Dublin. Thought it is an interesting and different pic to add.

Also went on a trip to South Korea and Japan. Both places are very very clean and it was almost impossible to find a place where to put a sticker, but managed to find one in Seoul and Tokyo! Kyoto and Osaka only have pictures of me holding a sticker.

Hope one day other people get the opportunity to visit these places and find those stickers <3